+++ gradle-sdk-project/asakusa-gradle-plugins/src/test/groovy/com/asakusafw/gradle/plugins/AsakusaUpgradeTest.groovy
@@ -15,10 +15,6 @@
  */
 package com.asakusafw.gradle.plugins
 
-import org.gradle.testkit.runner.BuildResult
-import org.gradle.testkit.runner.GradleRunner
-import org.gradle.testkit.runner.TaskOutcome
-import org.gradle.testkit.runner.UnexpectedBuildResultException
 import org.gradle.util.GradleVersion
 import org.junit.Rule
 import org.junit.Test
@@ -95,127 +91,8 @@ class AsakusaUpgradeTest {
     }
 
     private void doUpgrade(String version) {
-        Set<File> classpath = toClasspath(AsakusafwBasePlugin, 'META-INF/gradle-plugins/asakusafw-sdk.properties')
-        String script = """
-            buildscript {
-                repositories {
-                    mavenCentral()
-                }
-                dependencies {
-                    classpath files(${classpath.collect { "'''${it.absolutePath}'''" }.join(', ')})
-                    classpath 'org.codehaus.groovy:groovy-backports-compat23:${GroovySystem.version}'
-                }
-            }
-            apply plugin: 'asakusafw-sdk'
-            apply plugin: 'asakusafw-organizer'
-            """.stripIndent()
-        File buildScript = projectDir.newFile('build.gradle')
-        buildScript.setText(script, 'UTF-8')
-        BuildResult result
-        try {
-            result = GradleRunner.create()
-                    .withGradleVersion(version)
-                    .withProjectDir(projectDir.root)
-                    .withArguments(AsakusafwBasePlugin.TASK_UPGRADE, '-i', '-s')
-                    .build()
-        } catch (UnexpectedBuildResultException t) {
-            throw new AssertionError(script, t)
-        }
-        if (GradleVersion.version(version).compareTo(GradleVersion.version('2.5')) >= 0) {
-            assert result.task(":${AsakusafwBasePlugin.TASK_UPGRADE}").outcome == TaskOutcome.SUCCESS
-        }
-    }
-
-    private static Set<File> toClasspath(Object... resources) {
-        Set<File> results = new LinkedHashSet<>()
-        for (Object resource : resources) {
-            if (resource instanceof Class<?>) {
-                results.add(toClasspathEntry((Class<?>) resource))
-            } else if (resource instanceof String) {
-                results.add(toClasspathEntry((String) resource))
-            } else {
-                throw new AssertionError(resource)
-            }
-        }
-        return results
-    }
-
-    private static File toClasspathEntry(String resourcePath) {
-        URL resource = AsakusaUpgradeTest.class.getClassLoader().getResource(resourcePath)
-        if (resource == null) {
-            throw new AssertionError(resourcePath)
-        }
-        return findLibraryFromUrl(resource, resourcePath)
-    }
-
-    private static File toClasspathEntry(Class<?> aClass) {
-        URL resource = toUrl(aClass)
-        if (resource == null) {
-            throw new AssertionError(aClass)
-        }
-        String resourcePath = toResourcePath(aClass)
-        return findLibraryFromUrl(resource, resourcePath)
-    }
-
-    private static String toResourcePath(Class<?> aClass) {
-        return aClass.getName().replace('.', '/') + '.class'
-    }
-
-    private static URL toUrl(Class<?> aClass) {
-        String className = aClass.getName()
-        int start = className.lastIndexOf('.') + 1
-        String name = className.substring(start)
-        URL resource = aClass.getResource(name + '.class')
-        return resource
-    }
-
-    private static Set<File> findLibrariesByResource(ClassLoader classLoader, String path) {
-        Set<File> results = new LinkedHashSet<>()
-        for (URL url : Collections.list(classLoader.getResources(path))) {
-            File library = findLibraryFromUrl(url, path)
-            if (library != null) {
-                results.add(library)
-            }
-        }
-        return results
-    }
-
-    private static File findLibraryFromUrl(URL resource, String resourcePath) {
-        String protocol = resource.getProtocol()
-        if (protocol.equals('file')) {
-            File file = new File(resource.toURI())
-            return toClassPathRoot(file, resourcePath)
-        } else if (protocol.equals('jar')) {
-            String path = resource.getPath()
-            return toClassPathRoot(path, resourcePath)
-        } else {
-            throw new AssertionError(resource)
-        }
-    }
-
-    private static File toClassPathRoot(File resourceFile, String resourcePath) {
-        File current = resourceFile.getParentFile()
-        for (int start = resourcePath.indexOf('/'); start >= 0; start = resourcePath.indexOf('/', start + 1)) {
-            current = current.getParentFile()
-            if (current == null || current.isDirectory() == false) {
-                throw new AssertionError(resourceFile)
-            }
-        }
-        return current
-    }
-
-    private static File toClassPathRoot(String uriQualifiedPath, String resourceName) {
-        int entry = uriQualifiedPath.lastIndexOf('!')
-        String qualifier
-        if (entry >= 0) {
-            qualifier = uriQualifiedPath.substring(0, entry)
-        } else {
-            qualifier = uriQualifiedPath
-        }
-        URI archive = new URI(qualifier)
-        if (archive.getScheme().equals('file') == false) {
-            throw new AssertionError(archive)
-        }
-        return new File(archive)
+        Set<File> classpath = GradleTestkitHelper.toClasspath(AsakusafwBasePlugin, 'META-INF/gradle-plugins/asakusafw-sdk.properties')
+        String script = GradleTestkitHelper.getSimpleBuildScript(classpath, 'asakusafw-sdk', 'asakusafw-organizer')
+        GradleTestkitHelper.runGradle(projectDir.root, version, script, AsakusafwBasePlugin.TASK_UPGRADE)
     }
 }
